scheduler_trx.c: cast ptrdiff value to fix printf format
On an ARM toolchain:
scheduler_trx.c:294:3: warning: format '%ld' expects argument of type 'long int', but argument 10 has type 'int'

Let's cast it to long int to make sure correct size is applied in all
platforms.

+++ src/osmo-bts-trx/scheduler_trx.c
@@ -292,7 +292,7 @@ got_msg:
 	/* check validity of message */
 	if (rc < 0) {
 		LOGL1S(DL1P, LOGL_FATAL, l1t, tn, chan, fn, "Prim invalid length, please FIX! "
-			"(len=%ld)\n", msg->tail - msg->l2h);
+			"(len=%ld)\n", (long)(msg->tail - msg->l2h));
 		/* free message */
 		msgb_free(msg);
 		goto no_msg;
